requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pymysql==1.0.2
pydantic==1.8.2
python-dotenv==0.19.0
email-validator>=2.1.0
cryptography>=41.0.0
pytest==6.2.5
httpx==0.19.0
python-multipart==0.0.5 